extract fit text mechanism into an until function move and add appendFittedText test
