+++ src/graph/node/fitTextRenderer.js
@@ -1,69 +1,12 @@
 import * as d3 from "d3";
 import provider from "../../provider/provider";
+import appendFittedText from "../../graph/util/appendFittedText";
 
 var fitTextRenderer = {};
 
-var CONTEXT_2D = document.createElement("canvas").getContext("2d");
-var DEFAULT_CANVAS_LINE_HEIGHT = 12;
-
-fitTextRenderer.getNodeBoundingBox = function (node) {
+fitTextRenderer.getNodeBoundingBox = function(node) {
     return node.getBBox();
 };
-
-export function measureTextWidth(text) {
-    return CONTEXT_2D.measureText(text).width;
-}
-
-/**
- * Inspired by https://beta.observablehq.com/@mbostock/fit-text-to-circle
- * @param text
- * @returns {*}
- */
-export function getLines(text) {
-    if (text === undefined || text === null) {
-        return {lines: []}
-    }
-
-    var text = String(text);
-
-    var words = text.split(/\s+/g); // To hyphenate: /\s+|(?<=-)/
-    if (!words[words.length - 1]) words.pop();
-    if (!words[0]) words.shift();
-
-    var targetWidth = Math.sqrt(measureTextWidth(text.trim()) * DEFAULT_CANVAS_LINE_HEIGHT);
-
-    var line;
-    var lineWidth0 = Infinity;
-    var lines = [];
-
-    for (var i = 0, n = words.length; i < n; ++i) {
-        var lineText1 = (line ? line.text + " " : "") + words[i];
-        var lineWidth1 = measureTextWidth(lineText1);
-        if ((lineWidth0 + lineWidth1) / 2 < targetWidth) {
-            line.width = lineWidth0 = lineWidth1;
-            line.text = lineText1;
-        } else {
-            lineWidth0 = measureTextWidth(words[i]);
-            line = {width: lineWidth0, text: words[i]};
-            lines.push(line);
-        }
-    }
-
-    var textRadius = 0;
-
-    for (var i = 0, n = lines.length; i < n; ++i) {
-        var dy = (Math.abs(i - n / 2 + 0.5) + 0.5) * DEFAULT_CANVAS_LINE_HEIGHT;
-        var dx = lines[i].width / 2;
-        textRadius = Math.max(textRadius, Math.sqrt(dx * dx + dy * dy));
-    }
-
-    var lines2 = lines.map(function (d) {
-        return {"text": d.text, "linesLength": lines.length}
-    });
-
-    return {"lines": lines2, "textRadius": textRadius}
-}
-
 
 /**
  * Create the text representation of a node by slicing the text into lines to fit the node.
@@ -82,42 +25,10 @@ fitTextRenderer.render = function (nodeSelection) {
             return provider.node.getCSSClass(node, "back-text")
         });
 
-
-    var textMiddle = nodeSelection.append('text')
-        .attr("class", function (node) {
-            return provider.node.getCSSClass(node, "text")
-        });
-
-    // TODO extract style in CSS
-    textMiddle.attr("style", "text-anchor: middle; font: 10px sans-serif")
-        .datum(function (d) {
-            var text = provider.node.getTextValue(d);
-            var lines = getLines(text);
-            d.lines = lines.lines;
-            d.textRadius = lines.textRadius;
-            return d
-        });
-
-    textMiddle.selectAll("tspan")
-        .data(function (d) {
-            return d.lines
-        })
-        .enter().append("tspan")
-        .attr("x", 0)
-        .attr("y", function (d, i, nodes) {
-            return (i - d.linesLength / 2 + 0.8) * DEFAULT_CANVAS_LINE_HEIGHT;
-        })
-        .text(function (d) {
-            return d.text;
-        });
-
-    textMiddle.attr("transform", function (d) {
-        var scale = 1;
-        if (d.textRadius !== 0 && d.textRadius) {
-            scale = provider.node.getSize(d) / d.textRadius;
-        }
-        return "translate(" + 0 + "," + 0 + ")" + " scale(" + scale + ")"
-    });
+    appendFittedText(nodeSelection,
+        function (d) { return provider.node.getTextValue(d)},
+        function (d) { return provider.node.getSize(d) },
+        function (d) { return provider.node.getCSSClass(d, "text") });
 
     backgroundRectSelection
         .attr("x", function (d) {
